+++ DataForge.Core/src/DataForge.Core.Excel/ExcelPipelineExtensions.cs
@@ -1,3 +1,4 @@
+using DataForge.Core;
 using DataForge.Core.Core.Models;
 using DataForge.Core.Core.Pipeline;
 using System.Threading;
@@ -7,11 +8,14 @@ namespace DataForge.Core.Excel;
 
 public static class ExcelPipelineExtensions
 {
-    public static IDataPipeline<T> FromExcel<T>(this DataForgePipeline builder, string filePath, ExcelSourceOptions? options = null) where T : new()
+    public static IDataPipeline<T> FromExcel<T>(string filePath, ExcelSourceOptions? options = null) where T : new()
     {
         var source = new ExcelSource<T>(filePath, options);
         return new DataPipeline<T>(source.ReadAsync());
     }
+
+    public static IDataPipeline<T> FromExcel<T>(this DataForgePipeline _, string filePath, ExcelSourceOptions? options = null) where T : new()
+        => FromExcel<T>(filePath, options);
 
     public static Task<ExportResults> ToExcel<T>(this IDataPipeline<T> pipeline, string filePath, ExcelExportOptions? options = null, CancellationToken cancellationToken = default)
     {